# services/user_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for an e-commerce microservice with authentication and user management.",
        "title": "E-commerce Microservice API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.mit.edu/license"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/v1/api/users": {
            "get": {
                "description": "Retrieve a user based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/users/addresses": {
            "get": {
                "description": "Retrieves a list of addresses associated with the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Get user addresses",
                "responses": {
                    "200": {
                        "description": "Addresses retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new address for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Add user address",
                "parameters": [
                    {
                        "description": "Address details",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddressInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Address successfully created",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/users/addresses/{id}": {
            "put": {
                "description": "Updates an existing address for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Update user address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated address details",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddressInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address successfully updated",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing address for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete user address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddressInsertDTO": {
            "type": "object",
            "required": [
                "address_line_1",
                "city",
                "country",
                "postal_code",
                "state"
            ],
            "properties": {
                "address_line_1": {
                    "description": "The first line of the address, required with a length between 5 and 100 characters.\nExample: 123 Main Street\n@Param address_line_1 body string true \"Address Line 1\" validate:\"required,min=5,max=100\"",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 5
                },
                "address_line_2": {
                    "description": "The second line of the address, optional with a length between 5 and 100 characters.\nExample: Apt 4B\n@Param address_line_2 body string false \"Address Line 2 (Optional)\" validate:\"omitempty,min=5,max=100\"",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 5
                },
                "city": {
                    "description": "The city of the address, required with a length between 2 and 50 characters.\nExample: New York\n@Param city body string true \"City\" validate:\"required,min=2,max=50\"",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "country": {
                    "description": "The country code of the address, required in ISO 3166-1 alpha-2 format.\nExample: US\n@Param country body string true \"Country (ISO 3166-1 alpha-2)\" validate:\"required,len=2\"",
                    "type": "string"
                },
                "is_default": {
                    "description": "Whether this address is the default for the user.\nExample: true\n@Param is_default body bool true \"Is Default Address\"",
                    "type": "boolean"
                },
                "postal_code": {
                    "description": "The postal or ZIP code of the address, required.\nExample: 10001\n@Param postal_code body string true \"Postal Code\" validate:\"required\"",
                    "type": "string"
                },
                "state": {
                    "description": "The state or province of the address, required with a length between 2 and 50 characters.\nExample: NY\n@Param state body string true \"State/Province\" validate:\"required,min=2,max=50\"",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "user_id": {
                    "description": "The unique identifier of the user associated with this address.\nExample: 550e8400-e29b-41d4-a716-446655440000\n@Param user_id body string true \"User ID (UUID)\"",
                    "type": "string"
                }
            }
        },
...
        "response.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The HTTP status code of the response.\nExample: 200\n@Param code query int true \"HTTP status code\"",
                    "type": "integer"
                },
                "data": {
                    "description": "Data contains the response data, if any.\nExample: { \"id\": 123, \"name\": \"John Doe\" }\n@Param data query object false \"Response data\""
                },
                "errors": {
                    "description": "Errors contains error details, if any.\nExample: \"Invalid input data\"\n@Param errors query object false \"Error details\""
                },
                "message": {
                    "description": "A message that gives more information about the response.\nExample: \"Request successful\"\n@Param message query string true \"Response message\"",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates whether the request was successful or not.\nExample: true\n@Param success query bool true \"Success status\"",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "Timestamp represents the time when the response was generated.\nExample: \"2025-02-24T12:34:56Z\"\n@Param timestamp query string true \"Timestamp of the response\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}